+++ adminer/drivers/mysql.inc.php
@@ -1117,7 +1117,7 @@ WHERE ROUTINE_SCHEMA = DATABASE() AND ROUTINE_TYPE = '$type' AND ROUTINE_NAME = 
 	}
 
 	/** Check whether a feature is supported
-	* @param literal-string $feature check|comment|columns|copy|database|descidx|drop_col|dump|event|indexes|kill|materializedview
+	* @param literal-string $feature check|comment|columns|copy|cursor|database|descidx|drop_col|dump|event|indexes|kill|materializedview
 	* |move_col|privileges|procedure|processlist|routine|scheme|sequence|sql|status|table|transaction_ddl|trigger|type|variables|view|view_trigger
 	*/
 	function support(string $feature): bool {